refactor(trubill-vault): consolidate toBN helper into common/amounts
Why:
* toBN belongs with the amount converters it pairs with; a separate web3/bn module was redundant.

What changed:
* Moved toBN into common/amounts.ts and removed common/web3/bn.ts.
* Repointed the anchor and native instruction imports to common/amounts.
* Added section header comments in common/web3/pda.ts.

Impact:
* None (behavior unchanged).
