+++ packages/trubill-vault/anchor/instructions/deposit.ts
@@ -3,9 +3,8 @@ import { BN, type Program } from "@coral-xyz/anchor";
 import { ASSOCIATED_TOKEN_PROGRAM_ID, TOKEN_2022_PROGRAM_ID, TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import { PublicKey, SystemProgram } from "@solana/web3.js";
 import { USDC_MINT } from "../../common/addresses";
-import { usdc } from "../../common/amounts";
+import { toBN, usdc } from "../../common/amounts";
 import { type TrubillVault } from "../../common/idls/trubill_vault";
-import { toBN } from "../../common/web3/bn";
 import { getWalletKeypair } from "../../common/web3/env";
 import * as Pda from "../../common/web3/pda";
 import { deriveATAAddress } from "../../common/web3/token";
