+++ packages/trubill-vault/anchor/instructions/request-redeem.ts
@@ -2,9 +2,8 @@ import { fileURLToPath } from "node:url";
 import { BN, type Program } from "@coral-xyz/anchor";
 import { TOKEN_2022_PROGRAM_ID } from "@solana/spl-token";
 import { type PublicKey, SystemProgram } from "@solana/web3.js";
-import { trubill } from "../../common/amounts";
+import { toBN, trubill } from "../../common/amounts";
 import { type TrubillVault } from "../../common/idls/trubill_vault";
-import { toBN } from "../../common/web3/bn";
 import { getWalletKeypair } from "../../common/web3/env";
 import * as Pda from "../../common/web3/pda";
 import { deriveATAAddress } from "../../common/web3/token";
