+++ packages/trubill-vault/native/instructions/deposit.ts
@@ -3,8 +3,7 @@ import { BN } from "@coral-xyz/anchor";
 import { ASSOCIATED_TOKEN_PROGRAM_ID, TOKEN_2022_PROGRAM_ID, TOKEN_PROGRAM_ID } from "@solana/spl-token";
 import { PublicKey, SystemProgram, TransactionInstruction } from "@solana/web3.js";
 import { TRUBILL_VAULT_PROGRAM_ID, USDC_MINT } from "../../common/addresses";
-import { usdc } from "../../common/amounts";
-import { toBN } from "../../common/web3/bn";
+import { toBN, usdc } from "../../common/amounts";
 import { getConnection, getWalletKeypair } from "../../common/web3/env";
 import * as Pda from "../../common/web3/pda";
 import { deriveATAAddress } from "../../common/web3/token";
